+++ resteasy-spring/src/main/java/org/jboss/resteasy/plugins/spring/EmbeddedContainerBean.java
@@ -1,9 +1,9 @@
 package org.jboss.resteasy.plugins.spring;
 
+import org.jboss.resteasy.client.jaxrs.util.PortProvider;
 import org.jboss.resteasy.core.Dispatcher;
 import org.jboss.resteasy.plugins.server.embedded.SecurityDomain;
 import org.jboss.resteasy.plugins.server.tjws.TJWSEmbeddedJaxrsServer;
-import org.jboss.resteasy.test.TestPortProvider;
 import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.InitializingBean;
 
@@ -11,7 +11,7 @@ public class EmbeddedContainerBean implements InitializingBean, DisposableBean
 {
    private TJWSEmbeddedJaxrsServer embeddedServer = new TJWSEmbeddedJaxrsServer();
    private String rootPath = "/";
-   private int port = TestPortProvider.getPort();
+   private int port = PortProvider.getPort();
    private SecurityDomain securityDoamin = null;
 
    public String getRootPath()
